@@ -3,7 +3,7 @@
   "description": "Example app for the ngx-swiper-wrapper",
   "private": true,
   "license": "MIT",
-  "version": "6.0.0",
+  "version": "8.0.0",
   "tslint": {
     "extends": "./tslint.json"
   },
@@ -12,23 +12,18 @@
   },
   "scripts": {
     "start": "ng serve",
-    "clean": "rm -rf dist",
     "lint": "npm run lint:ts && npm run lint:css",
-    "build": "rm -rf dist/* && npm run build:aot",
+    "build": "ng build --configuration=production",
     "deploy": "rm -rf dist && deploy-to-git",
     "lint:ts": "tslint -p src/tsconfig.json \"src/**/*.ts\"",
-    "lint:css": "stylelint --syntax scss \"src/**/*.css\"",
-    "build:aot": "webpack --config config/webpack-build-aot.config.js",
-    "build:jit": "webpack --config config/webpack-build-jit.config.js",
-    "serve:sjs": "lite-server -c config/lite-server-bs-config.js",
-    "serve:wds": "webpack-dev-server --open --inline --config config/webpack-server-jit.config.js"
+    "lint:css": "stylelint --syntax scss \"src/**/*.css\""
   },
   "config": {
     "deployToGit": {
       "repository": "git@github.com:zefoy/ngx-swiper-wrapper.git",
       "branch": "gh-pages",
       "folder": "dist",
-      "script": "npm run build",
+      "script": "npm run build -- --base-href=ngx-swiper-wrapper --delete-output-path=false",
       "commit": "Publishing $npm_package_version",
       "user": {
         "name": "ZEF Devel",
@@ -37,46 +32,32 @@
     }
   },
   "dependencies": {
-    "@angular/cdk": "^6.0.0",
-    "@angular/common": "^6.0.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/core": "^6.0.0",
-    "@angular/flex-layout": "6.0.0-beta.16",
-    "@angular/platform-browser": "^6.0.0",
-    "@angular/platform-browser-dynamic": "^6.0.0",
-    "core-js": "^2.5.0",
-    "ngx-swiper-wrapper": "^6.0.0",
-    "rxjs": "^6.2.0",
-    "swiper": "^4.3.0",
-    "zone.js": "^0.8.0"
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/flex-layout": "8.0.0-beta.26",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "core-js": "^3.1.0",
+    "ngx-swiper-wrapper": "^8.0.0",
+    "rxjs": "^6.5.0",
+    "swiper": "^4.5.0",
+    "zone.js": "^0.9.0"
   },
   "devDependencies": {
-    "@angular/cli": "^6.0.0",
-    "@angular/compiler-cli": "^6.0.0",
-    "@angular-devkit/build-ng-packagr": "^0.6.0",
-    "@angular-devkit/build-angular": "^0.6.0",
-    "@ngtools/webpack": "^6.0.0",
-    "@types/node": "^8.10.0",
-    "@types/swiper": "^4.2.0",
-    "codelyzer": "^4.3.0",
-    "connect-history-api-fallback": "^1.5.0",
-    "copy-webpack-plugin": "^4.5.0",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@ngtools/webpack": "^8.0.0",
+    "@types/node": "^12.0.0",
+    "@types/swiper": "^4.4.0",
+    "codelyzer": "^5.1.0",
     "deploy-to-git": "^0.1.0",
-    "html-webpack-plugin": "^3.2.0",
-    "lite-server": "^2.3.0",
-    "ng-packagr": "^3.0.0",
-    "node-sass": "^4.9.0",
-    "plugin-typescript": "^8.0.0",
-    "raw-loader": "^0.5.0",
-    "sass-loader": "^7.0.0",
-    "stylelint": "^9.3.0",
-    "stylelint-config-standard": "^18.2.0",
-    "stylelint-order": "^0.8.0",
-    "systemjs": "^0.21.0",
-    "tslint": "^5.10.0",
-    "typescript": "~2.7.0",
-    "webpack": "^4.12.0",
-    "webpack-cli": "^3.0.0",
-    "webpack-dev-server": "^3.1.0"
+    "stylelint": "^10.1.0",
+    "stylelint-config-standard": "^18.3.0",
+    "stylelint-order": "^3.0.0",
+    "tslint": "^5.17.0",
+    "typescript": "~3.4.0"
   }
 }
