@@ -3,7 +3,7 @@
   "description": "Angular wrapper library for Swiper",
   "bugs": "https://github.com/zefoy/ngx-swiper-wrapper/issues",
   "license": "MIT",
-  "version": "6.3.0",
+  "version": "8.0.2",
   "main": "./bundles/ngx-swiper-wrapper.umd.js",
   "module": "./dist/ngx-swiper-wrapper.es5.js",
   "typings": "./dist/ngx-swiper-wrapper.d.ts",
@@ -17,54 +17,52 @@
     "clean": "rm -rf dist bundles",
     "start": "npm run build && npm run bundle",
     "watch": "watch --wait=5 'npm run start' src/",
-    "lint": "npm run lint:ts && npm run lint:scss",
+    "lint": "npm run lint:ts && npm run lint:css",
     "build": "npm run build:html && npm run build:css && npm run build:js",
     "bundle": "npm run bundle:es5 ; npm run bundle:umd ; npm run minify:umd",
     "prepare": "npm run clean && npm run build && npm run bundle",
-    "lint:ts": "tslint -p src/tsconfig.json \"src/**/*.ts\"",
-    "lint:scss": "stylelint --syntax scss \"src/**/*.scss\"",
+    "lint:ts": "tslint \"src/**/*.ts\"",
+    "lint:css": "stylelint \"src/**/*.css\"",
     "build:js": "ngc -p src/tsconfig.json",
-    "build:css": "node-sass src/ -o dist/ --source-map=true",
-    "build:html": "cpx --dereference src/lib/*.html dist/lib/",
+    "build:css": "cpx src/lib/swiper.component.css dist/lib/",
+    "build:html": "cpx src/lib/swiper.component.html dist/lib/",
     "bundle:es5": "rollup -c config/rollup-es5.config.js --silent",
     "bundle:umd": "rollup -c config/rollup-umd.config.js --silent",
-    "minify:umd": "uglifyjs bundles/$npm_package_name.umd.js -o bundles/$npm_package_name.umd.min.js --source-map=\"filename=bundles/$npm_package_name.umd.min.js.map\""
+    "minify:umd": "terser bundles/$npm_package_name.umd.js -o bundles/$npm_package_name.umd.min.js --source-map=\"filename=bundles/$npm_package_name.umd.min.js.map\""
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/zefoy/ngx-swiper-wrapper.git"
   },
   "dependencies": {
-    "swiper": "^4.3.0"
+    "swiper": "^4.5.0"
   },
   "devDependencies": {
-    "@angular/cli": "^6.0.0",
-    "@angular/common": "^6.0.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/compiler-cli": "^6.0.0",
-    "@angular/core": "^6.0.0",
-    "@angular-devkit/build-ng-packagr": "^0.6.0",
-    "@angular-devkit/build-angular": "^0.6.0",
-    "@angular/platform-browser": "^6.0.0",
-    "@types/node": "^8.10.0",
-    "@types/swiper": "^4.2.0",
-    "codelyzer": "^4.3.0",
-    "core-js": "^2.5.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@types/node": "^12.0.0",
+    "@types/swiper": "^4.4.0",
+    "codelyzer": "^5.1.0",
+    "core-js": "^3.1.0",
     "cpx": "^1.5.0",
-    "ng-packagr": "^3.0.0",
-    "node-sass": "^4.9.0",
-    "rollup": "^0.60.0",
-    "rollup-plugin-node-resolve": "^3.3.0",
+    "rollup": "^1.14.0",
+    "rollup-plugin-node-resolve": "^5.0.0",
     "rollup-plugin-sourcemaps": "^0.4.0",
-    "rxjs": "^6.2.0",
-    "stylelint": "^9.3.0",
-    "stylelint-config-standard": "^18.2.0",
-    "stylelint-order": "^0.8.0",
-    "tslint": "^5.10.0",
-    "typescript": "~2.7.0",
-    "uglify-es": "^3.3.0",
+    "rxjs": "^6.5.0",
+    "stylelint": "^10.1.0",
+    "stylelint-config-standard": "^18.3.0",
+    "stylelint-order": "^3.0.0",
+    "terser": "^4.0.0",
+    "tsickle": "^0.35.0",
+    "tslint": "^5.17.0",
+    "typescript": "~3.4.0",
     "watch": "^1.0.0",
-    "zone.js": "^0.8.0"
+    "zone.js": "^0.9.0"
   },
   "peerDependencies": {
     "@angular/common": ">=5.0.0",
